introduce ITargetFinder
