+++ jsettlers.logic/src/main/java/jsettlers/ai/highlevel/pioneers/PioneerGroup.java
@@ -12,8 +12,9 @@
  * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  *******************************************************************************/
-package jsettlers.ai.highlevel;
+package jsettlers.ai.highlevel.pioneers;
 
+import jsettlers.ai.highlevel.AiStatistics;
 import jsettlers.common.movable.EMovableAction;
 import jsettlers.common.movable.EMovableType;
 import jsettlers.common.position.ShortPoint2D;
